Move API policies to MediaBrowser.Common
