+++ Jellyfin.Api/Controllers/LocalizationController.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using Jellyfin.Api.Constants;
+using MediaBrowser.Common.Api;
 using MediaBrowser.Model.Entities;
 using MediaBrowser.Model.Globalization;
 using Microsoft.AspNetCore.Authorization;
