+++ Jellyfin.Server/Extensions/ApiServiceCollectionExtensions.cs
@@ -22,6 +22,7 @@ using Jellyfin.Data.Enums;
 using Jellyfin.Extensions.Json;
 using Jellyfin.Server.Configuration;
 using Jellyfin.Server.Filters;
+using MediaBrowser.Common.Api;
 using MediaBrowser.Common.Net;
 using MediaBrowser.Model.Entities;
 using MediaBrowser.Model.Session;
